+++ src/main/java/ru/kappers/logic/contract/IContract.java
@@ -7,8 +7,8 @@ import java.util.Map;
 public interface IContract {
     KapperInfo initCapper(User user);
     KapperInfo getKapperInfo(User user);
-    void blockTokens(User user, double amount);
-    void unblockAmount(User user, double amount);
-    void withdrawBlockedTokens(User user, double amount);
+    void blockTokens(User user, Integer amount);
+    void unblockAmount(User user, Integer amount);
+    void withdrawBlockedTokens(User user, Integer amount);
     Map<User, KapperInfo> getAllInfo();
 }
